backend/requirements.txt
flask
python-dotenv
flask-cors
requests
streamlit
pandas
matplotlib
plotly
